requirements.txt
cryptography==2.9.2
Django==3.0.5
jsonpath==0.82
APScheduler==3.6.3
pytz==2019.3
PyMySQL==0.9.3
requests==2.24.0
django_redis==4.11.0
pipreqs==0.4.10
djangorestframework==3.11.0
xlrd==1.2.0
pymssql==2.1.5
gunicorn==20.0.4
xlwt==1.3.0
cx-Oracle==8.0.1
django-cors-headers==3.5.0
django-mysql==3.10.0
PyJWT==1.7.1



